Update sentry.proto interface; closes #481
