+++ interfaces/sentry.proto
@@ -115,21 +115,51 @@ enum Protocol {
   ETH66 = 1;
 }
 
-message SetStatusReply {
-  Protocol protocol = 1;
-}
+message SetStatusReply {}
+
+message HandShakeReply {Protocol protocol = 1;}
 
 message MessagesRequest {
   repeated MessageId ids = 1;
 }
 
+message PeerCountRequest {}
+
+message PeerCountReply {uint64 count = 1;}
+
+message PeersRequest {}
+message PeersReply {
+  enum PeerEvent {
+    Connect = 0;
+    Disconnect = 1;
+  }
+  types.H256 peer_id = 1;
+  PeerEvent event = 2;
+}
+
 service Sentry {
+  // SetStatus - force new ETH client state of sentry - network_id, max_block, etc...
+  rpc SetStatus(StatusData) returns (SetStatusReply);
+
   rpc PenalizePeer(PenalizePeerRequest) returns (google.protobuf.Empty);
   rpc PeerMinBlock(PeerMinBlockRequest) returns (google.protobuf.Empty);
+
+  // HandShake - pre-requirement for all Send* methods - returns ETH protocol version,
+  // without knowledge of protocol - impossible encode correct P2P message
+  rpc HandShake(google.protobuf.Empty) returns (HandShakeReply);
   rpc SendMessageByMinBlock(SendMessageByMinBlockRequest) returns (SentPeers);
   rpc SendMessageById(SendMessageByIdRequest) returns (SentPeers);
   rpc SendMessageToRandomPeers(SendMessageToRandomPeersRequest) returns (SentPeers);
   rpc SendMessageToAll(OutboundMessageData) returns (SentPeers);
-  rpc SetStatus(StatusData) returns (SetStatusReply);
+
+  // Subscribe to receive messages.
+  // Calling multiple times with a different set of ids starts separate streams.
+  // It is possible to subscribe to the same set if ids more than once.
   rpc Messages(MessagesRequest) returns (stream InboundMessage);
+  rpc PeerCount(PeerCountRequest) returns (PeerCountReply);
+
+  // Notifications about connected (after sub-protocol handshake) or lost peer
+  rpc Peers(PeersRequest) returns (stream PeersReply);
+  // NodeInfo returns a collection of metadata known about the host.
+  rpc NodeInfo(google.protobuf.Empty) returns(types.NodeInfoReply);
 }
