+++ node/silkworm/downloader/messages/InboundGetBlockBodies.cpp
@@ -19,7 +19,7 @@
 #include <silkworm/common/log.hpp>
 #include <silkworm/downloader/internals/body_retrieval.hpp>
 #include <silkworm/downloader/packets/BlockBodiesPacket.hpp>
-#include <silkworm/downloader/rpc/SendMessageById.hpp>
+#include <silkworm/downloader/rpc/send_message_by_id.hpp>
 
 namespace silkworm {
