+++ node/silkworm/downloader/messages/InboundNewBlock.cpp
@@ -21,7 +21,7 @@
 #include <silkworm/common/cast.hpp>
 #include <silkworm/common/log.hpp>
 #include <silkworm/downloader/internals/random_number.hpp>
-#include <silkworm/downloader/rpc/SendMessageById.hpp>
+#include <silkworm/downloader/rpc/send_message_by_id.hpp>
 
 namespace silkworm {
