+++ node/silkworm/downloader/messages/InboundNewBlockHashes.cpp
@@ -21,7 +21,7 @@
 #include <silkworm/common/log.hpp>
 #include <silkworm/downloader/internals/random_number.hpp>
 #include <silkworm/downloader/packets/RLPEth66PacketCoding.hpp>
-#include <silkworm/downloader/rpc/SendMessageById.hpp>
+#include <silkworm/downloader/rpc/send_message_by_id.hpp>
 
 namespace silkworm {
