+++ node/silkworm/downloader/rpc/hand_shake.cpp
@@ -14,14 +14,13 @@
    limitations under the License.
 */
 
-#include "SendMessageById.hpp"
+#include "hand_shake.hpp"
 
 namespace silkworm::rpc {
 
-SendMessageById::SendMessageById(const std::string& peerId, std::unique_ptr<sentry::OutboundMessageData> message)
-    : UnaryCall("SendMessageById", &sentry::Sentry::Stub::SendMessageById, {}) {
-    request_.set_allocated_peer_id(to_H512(peerId).release());
-    request_.set_allocated_data(message.release());  // take ownership
+HandShake::HandShake() : UnaryCall("HandShake", &sentry::Sentry::Stub::HandShake, {}) {
 }
 
 }  // namespace silkworm::rpc
+
+
