+++ node/silkworm/downloader/rpc/penalize_peer.cpp
@@ -14,17 +14,16 @@
    limitations under the License.
 */
 
-#ifndef SILKWORM_SENDMESSAGEBYID_HPP
-#define SILKWORM_SENDMESSAGEBYID_HPP
-
-#include <silkworm/downloader/sentry_client.hpp>
+#include "penalize_peer.hpp"
 
 namespace silkworm::rpc {
 
-class SendMessageById : public rpc::UnaryCall<sentry::Sentry, sentry::SendMessageByIdRequest, sentry::SentPeers> {
-  public:
-    SendMessageById(const std::string& peerId, std::unique_ptr<sentry::OutboundMessageData> message);
-};
+PenalizePeer::PenalizePeer(const std::string& peerId, Penalty penalty)
+    : UnaryCall("PenalizePeer", &sentry::Sentry::Stub::PenalizePeer, {}) {
+    request_.set_allocated_peer_id(to_H512(peerId).release());
+
+    sentry::PenaltyKind raw_penalty = static_cast<sentry::PenaltyKind>(penalty);
+    request_.set_penalty(raw_penalty);
+}
 
 }  // namespace silkworm::rpc
-#endif  // SILKWORM_SENDMESSAGEBYID_HPP
