+++ node/silkworm/downloader/rpc/send_message_to_all.cpp
@@ -1,7 +1,7 @@
 /*
    Copyright 2021 The Silkworm Authors
 
-   Licensed under the Apache License, Version 2.0 (the "License");
+Licensed under the Apache License, Version 2.0 (the "License");
    you may not use this file except in compliance with the License.
    You may obtain a copy of the License at
 
@@ -14,18 +14,13 @@
    limitations under the License.
 */
 
-#ifndef SILKWORM_PEERMINBLOCK_HPP
-#define SILKWORM_PEERMINBLOCK_HPP
-
-#include <silkworm/downloader/sentry_client.hpp>
+#include "send_message_to_all.hpp"
 
 namespace silkworm::rpc {
 
-class PeerMinBlock : public rpc::UnaryCall<sentry::Sentry, sentry::PeerMinBlockRequest, google::protobuf::Empty> {
-  public:
-    PeerMinBlock(const std::string& peerId, BlockNum minBlock);
-};
+SendMessageToAll::SendMessageToAll(std::unique_ptr<sentry::OutboundMessageData> message)
+    : UnaryCall("SendMessageToAll", &sentry::Sentry::Stub::SendMessageToAll, {}) {
+    request_.Swap(message.get());
+}
 
 }  // namespace silkworm::rpc
-
-#endif  // SILKWORM_PEERMINBLOCK_HPP
