+++ node/silkworm/downloader/rpc/send_message_to_all.hpp
@@ -14,8 +14,8 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-#ifndef SILKWORM_SENDMESSAGETOALL_HPP
-#define SILKWORM_SENDMESSAGETOALL_HPP
+#ifndef SILKWORM_SEND_MESSAGE_TO_ALL_HPP
+#define SILKWORM_SEND_MESSAGE_TO_ALL_HPP
 
 #include <silkworm/downloader/sentry_client.hpp>
 
@@ -28,4 +28,4 @@ class SendMessageToAll : public rpc::UnaryCall<sentry::Sentry, sentry::OutboundM
 
 }  // namespace silkworm::rpc
 
-#endif  // SILKWORM_SENDMESSAGETOALL_HPP
+#endif  // SILKWORM_SEND_MESSAGE_TO_ALL_HPP
